- Restructrued everything kay gubot kaayo mo mo code.
