+++ Simple LMS OOP Project 2025/src/Dashboard/resources/courses_white.png
